+++ format.sh
@@ -1,0 +1,2 @@
+#!/bin/bash
+find . -regex '.*\.\(cpp\|hpp\)' -exec clang-format -style=file -i {} \;
