+++ src/sauros/processor/processor_load_lib.cpp
@@ -3,7 +3,7 @@
 #include <filesystem>
 #include <iostream>
 
-#include "../RLL/RLL.hpp"
+#include "../RLL/RLL.h"
 
 namespace sauros {
